+++ cypress/support/cypress-support/src/utils.js
@@ -16,7 +16,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 const CONSTANTS = require('../../constants/constants');
-const logger = require('../../Logger')('utils.js');
 const { _ } = Cypress;
 import { apiObject } from '../../appObjectConfigs';
 const MESSAGE = 'message';
@@ -24,6 +23,7 @@ const Validator = require('jsonschema').Validator;
 const validator = new Validator();
 const jsonFile = CONSTANTS.JSON_FILE_EXTENSION;
 let clientCreated = false;
+const { fireLog } = require('./fireLog');
 
 /**
  * @module utils
@@ -246,7 +246,7 @@ function getCommunicationMode() {
  * skipCurrentTest()
  */
 function skipCurrentTest() {
-  fireLog.info('The current test has been intentionally skipped by the test runner');
+  fireLog.info('The current test has been intentionally skipped by the test runner', 'report');
   mocha.suite.ctx.test?.skip();
 }
 
@@ -313,7 +313,7 @@ function getApiOrEventObjectFromGlobalList(method, context, appId, validationTyp
 
   // Failing when the filteredObjectList is empty.
   if (filteredObjectList.length < 1) {
-    fireLog.info('Could not find the api response in api list');
+    fireLog.info('Could not find the api response in api list', 'report');
     fireLog.isNotEmpty(filteredObjectList, 'filteredObjectList is not to be empty');
   }
 
@@ -332,7 +332,7 @@ function getApiOrEventObjectFromGlobalList(method, context, appId, validationTyp
 
   // If no response is found, fail with no appObject found.
   if (!extractedObject) {
-    fireLog.info(CONSTANTS.NO_APP_OR_EVENT_OBJECT);
+    fireLog.info(CONSTANTS.NO_APP_OR_EVENT_OBJECT, 'report');
     fireLog.assert(false, CONSTANTS.NO_APP_OR_EVENT_OBJECT);
   }
   return extractedObject;
@@ -352,7 +352,7 @@ function unsubscribe(webSocketClient = null) {
     throw new Error('Websocket client not established');
   }
   webSocketClient.unsubscribe(MESSAGE);
-  logger.info('Websocket connection closed Successfully', 'unsubscribe');
+  fireLog.info('Websocket connection closed Successfully');
 }
 
 /**
@@ -419,9 +419,9 @@ function getEnvVariable(variable, isRequired = true) {
 
   if (isRequired) {
     const errorMessage = `Required environment variable "${variable}" is missing or undefined.`;
-    logger.error(errorMessage, 'getEnvVariable');
+    fireLog.error(errorMessage);
     // To include stackTrace in the console
-    logger.error(stackTrace());
+    fireLog.error(stackTrace());
     throw new Error(errorMessage);
   }
   return envValue;
@@ -686,7 +686,7 @@ function checkForSecondaryAppId(appId) {
       return appId;
     }
   } catch (err) {
-    fireLog.info(eval(CONSTANTS.SECONDARY_APPID_MISSING_ERROR)).then(() => {
+    fireLog.info(eval(CONSTANTS.SECONDARY_APPID_MISSING_ERROR), 'report').then(() => {
       throw new Error(eval(CONSTANTS.SECONDARY_APPID_MISSING_ERROR));
     });
   }
@@ -703,176 +703,12 @@ function checkForSecondaryAppId(appId) {
 global.resolveDeviceVariable = function (key) {
   const resolvedDeviceData = Cypress.env('resolvedDeviceData');
   if (!(key in resolvedDeviceData)) {
-    logger.error(`Key ${key} not found in preprocessed data.`);
+    fireLog.error(`Key ${key} not found in preprocessed data.`);
     return null;
   }
-  logger.debug(`Resolved value for key ${key} is ${resolvedDeviceData[key]}`);
+  fireLog.debug(`Resolved value for key ${key} is ${resolvedDeviceData[key]}`);
   return resolvedDeviceData[key];
 };
-
-/**
- * FireLog class provides assertion methods with logging using Cypress's cy.log().
- * It wraps Cypress's assertion methods, allowing logging of messages for each assertion.
- * @class
- *
- * @example
- * // Usage example
- * fireLog.isNotNull(someValue, "Some message");
- * fireLog.isTrue(isTrueValue, "True message");
- * fireLog.isFalse(isFalseValue, "False message");
- * fireLog.deepEqual(actual, expected, "deepEqual message");
- *
- * fireLog.info('Discovery launch intent: ' + JSON.stringify(parsedIntent));
- * fireLog.info() is being used to log the message without any assertion.
- * Removing cy.log and replacing with fireLog.info() to get a cleaner report.
- *
- *
- */
-
-class FireLog extends Function {
-  constructor() {
-    // Creating the function body dynamically
-    const functionBody = `
-      return function (...args) {
-        return this.log(...args);
-      }
-    `;
-    super('...args', functionBody);
-
-    const handler = {
-      apply: function (target, thisArg, argumentsList) {
-        let message;
-        const methodName = target.name;
-        if (target.hasOwnLog) {
-          // If the method has its own logging, just apply it
-          return Reflect.apply(target, thisArg, argumentsList);
-        } else {
-          if (argumentsList.length > 3)
-            message =
-              'Expected: ' +
-              JSON.stringify(argumentsList[0]) +
-              ' Actual: ' +
-              'Expected : ' +
-              JSON.stringify(argumentsList[2]) +
-              ' Actual : ' +
-              JSON.stringify(argumentsList[1]);
-          else if (argumentsList.length == 3) message = argumentsList[2];
-          else if (argumentsList.length == 1) message = argumentsList[0];
-          else if (argumentsList.length == 2) message = argumentsList[1];
-          else
-            message =
-              argumentsList[argumentsList.length - 1] +
-              ' Actual: ' +
-              JSON.stringify(argumentsList[0]);
-          return cy.log(message).then(() => {
-            return Reflect.apply(target, thisArg, argumentsList);
-          });
-        }
-      },
-    };
-    // Proxy for the fireLog method
-    const instanceProxy = new Proxy(this, handler);
-    const fireLogProxy = new Proxy(instanceProxy, {
-      apply: function (target, thisArg, argumentsList) {
-        const message = argumentsList[argumentsList.length - 1];
-        return cy.log(message);
-      },
-    });
-
-    // Use cy.log(message) for every method in the class
-    const prototype = Object.getPrototypeOf(instanceProxy);
-    Object.getOwnPropertyNames(prototype).forEach((method) => {
-      if (
-        method !== 'constructor' &&
-        method !== 'fireLog' &&
-        typeof instanceProxy[method] === 'function'
-      ) {
-        instanceProxy[method] = new Proxy(instanceProxy[method], handler);
-        const methodSource = instanceProxy[method].toString();
-        instanceProxy[method].hasOwnLog = methodSource.includes('cy.log');
-      }
-    });
-
-    return fireLogProxy;
-  }
-
-  // Method to log a message without any assertion
-  log(message) {
-    return cy.log(message);
-  }
-
-  isNull(value, message) {
-    assert.isNull(value, message);
-  }
-
-  isNotNull(value, message) {
-    assert.isNotNull(value, message);
-  }
-
-  isUndefined(value, message) {
-    assert.isUndefined(value, message);
-  }
-
-  isTrue(value, message) {
-    assert.isTrue(value, message);
-  }
-
-  isFalse(value, message) {
-    assert.isFalse(value, message);
-  }
-
-  isOk(value, message) {
-    assert.isOk(value, message);
-  }
-
-  isNotEmpty(object, message) {
-    assert.isNotEmpty(object, message);
-  }
-
-  isBoolean(value, message) {
-    assert.isBoolean(value, message);
-  }
-
-  deepEqual(actual, expected, message) {
-    assert.deepEqual(actual, expected, message);
-  }
-
-  equal(actual, expected, message) {
-    assert.equal(actual, expected, message);
-  }
-
-  strictEqual(actual, expected, message) {
-    assert.strictEqual(actual, expected, message);
-  }
-
-  include(haystack, needle, message) {
-    cy.log(
-      message + ' ' + JSON.stringify(needle) + ' expected to be in ' + JSON.stringify(haystack)
-    );
-    assert.include(haystack, needle, message);
-  }
-  exists(value, message) {
-    assert.exists(value, message);
-  }
-
-  assert(expression, message) {
-    assert(expression, message);
-  }
-
-  fail(message) {
-    cy.log(message);
-    assert.fail(message);
-  }
-
-  info(message) {}
-
-  error(message) {
-    throw new Error(message);
-  }
-}
-
-const fireLog = new FireLog();
-global.fireLog = fireLog;
 
 /**
  * @module utils
@@ -923,7 +759,7 @@ global.extractEnvValue = function (attribute) {
   // Get the device data from env variable
   const deviceData = Cypress.env(CONSTANTS.DEVICE_DATA);
   if (!deviceData) {
-    logger.info('deviceData environment variable is not found');
+    fireLog.info('deviceData environment variable is not found');
   }
 
   // If the attribute starts with 'CYPRESSENV', extract nested property from env variable.
@@ -943,7 +779,7 @@ global.extractEnvValue = function (attribute) {
     if (envValue !== undefined) {
       attribute = envValue;
     } else {
-      logger.info(`Cypress env variable '${attribute}' does not exist`);
+      fireLog.info(`Cypress env variable '${attribute}' does not exist`);
     }
   }
   // Return the extracted value from device data or environment variable
@@ -1072,7 +908,7 @@ global.resolveAtRuntime = function (input) {
         element.includes('{{') ? replacingPatternOccurrenceWithValue(element) : element
       );
     } else {
-      logger.info(`Passed input - ${input} must be an array or a string.`);
+      fireLog.info(`Passed input - ${input} must be an array or a string.`);
     }
   };
 };
@@ -1191,12 +1027,12 @@ function applyOverrides(fireboltCallObject) {
 
     for (const override of overrides) {
       if (typeof override.applyWhen !== 'function') {
-        fireLog.info('Ignoring override: Missing applyWhen() function', override);
+        fireLog.info('Ignoring override: Missing applyWhen() function', 'report');
         continue;
       }
 
       if (!override.applyWhen()) {
-        fireLog.info('Ignoring override: applyWhen() returned false', override);
+        fireLog.info('Ignoring override: applyWhen() returned false', 'report');
         continue;
       }
       // Appending Override content to the fireboltCallObject if applyWhen() returns true
@@ -1205,7 +1041,7 @@ function applyOverrides(fireboltCallObject) {
   } catch (error) {
     fireLog.info(
       'Error in applyOverrides - Override key in the fireboltObject was not as expected::',
-      error
+      'report'
     );
   }
   return fireboltCallObject; // Return the original or modified object based on the override
@@ -1246,11 +1082,14 @@ function captureScreenshot() {
       method: method,
       params: param,
     };
-    fireLog.info(`Sending request to capture screenshot: ${JSON.stringify(screenshotRequest)}`);
+    fireLog.info(
+      `Sending request to capture screenshot: ${JSON.stringify(screenshotRequest)}`,
+      'report'
+    );
 
     try {
       cy.sendMessagetoPlatforms(screenshotRequest, 70000).then((response) => {
-        fireLog.info(`Screenshot capture response: ${JSON.stringify(response)}`);
+        fireLog.info(`Screenshot capture response: ${JSON.stringify(response)}`, 'report');
 
         const apiResponse = {
           response: response,
@@ -1309,7 +1148,6 @@ module.exports = {
   destroyGlobalObjects,
   writeJsonToFileForReporting,
   checkForTags,
-  fireLog,
   parseValue,
   checkForSecondaryAppId,
   resolveRecursiveValues,